services/regime_detection/requirements.txt
# Probabilistic Regime Detection Service - Dependencies
# Agent: regime-detection-specialist

# Core ML/Data Science
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.0

# Caching (optional, for future)
redis>=5.0.0

# Kafka (for event streaming)
kafka-python>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
httpx>=0.25.0  # For FastAPI test client

# Utilities
python-multipart>=0.0.6  # For FastAPI file uploads
python-dotenv>=1.0.0  # For environment variables
